+++ gdsfactory/simulation/gmeep/write_sparameters_meep.py
@@ -27,7 +27,7 @@ from gdsfactory.simulation.gmeep.get_simulation import (
     settings_get_simulation,
 )
 from gdsfactory.tech import LAYER_STACK, LayerStack
-from gdsfactory.types import PathType, PortSymmetries
+from gdsfactory.types import ComponentOrFactory, PathType, PortSymmetries
 
 ncores = multiprocessing.cpu_count()
 
@@ -129,7 +129,7 @@ def parse_port_eigenmode_coeff(port_index: int, ports, sim_dict: Dict):
 
 @pydantic.validate_arguments
 def write_sparameters_meep(
-    component: Component,
+    component: ComponentOrFactory,
     port_symmetries: Optional[PortSymmetries] = None,
     resolution: int = 30,
     wavelength_start: float = 1.5,
@@ -263,6 +263,9 @@ def write_sparameters_meep(
             where `a` is the angle in radians and `m` the module
 
     """
+    component = component() if callable(component) else component
+    assert isinstance(component, Component)
+
     for setting in settings.keys():
         if setting not in settings_get_simulation:
             raise ValueError(f"{setting} not in {settings_get_simulation}")